+++ src/Card/CardNumber/components/CardNumberInputs/CardNumberInputs.tsx
@@ -1,8 +1,8 @@
 import styles from "./CardNumberInputs.module.css";
-import Label from "../Label/Label";
-import Input from "../Input/Input";
-import { isAnyTrue } from "../../utils/isAnyTrue";
-import type { CardNumberInputKey, CardNumberState } from "../../constants/card";
+import Label from "../../../../components/Label/Label";
+import Input from "../../../../components/Input/Input";
+import { isAnyTrue } from "../../../../utils/isAnyTrue";
+import type { CardNumberInputKey, CardNumberState } from "../../types";
 
 export interface CardNumberInputsProps {
   cardNumberState: CardNumberState;
