+++ src/Card/CardNumber/constants/index.ts
@@ -1,0 +1,6 @@
+export const INITIAL_CARD_NUMBER_STATE = {
+  first: { value: "", isError: false },
+  second: { value: "", isError: false },
+  third: { value: "", isError: false },
+  fourth: { value: "", isError: false },
+} as const;
