--- src/components/CardInputBox/CardInputBox.module.css
@@ -1,14 +1,0 @@
-.title {
-  margin-bottom: 0.5rem;
-  color: #000;
-  font-family: "Noto Sans KR";
-  font-size: 1.5rem;
-  font-weight: 700;
-}
-
-.guideText {
-  margin-bottom: 1rem;
-  color: #8b95a1;
-  font-family: "Noto Sans KR";
-  font-size: 0.9rem;
-}
